refactor(scroll-engine): split the verb vocabulary into engine_verbs.cpp
Pure move of the P_SCROLL_VERB family (16 one-line verbs plus the two
hand-expanded ones) out of engine_navigation.cpp, which was 907 lines and
about to grow a cross-monitor focus arm. P_SCROLL_RESOLVE is duplicated in
the new TU; both TUs #undef their macros at end of file.
