+++ libs/phosphor-scroll-engine/CMakeLists.txt
@@ -41,6 +41,7 @@ add_library(PhosphorScrollEngine SHARED
     src/scrollengine/engine_lifecycle.cpp
     src/scrollengine/engine_float.cpp
     src/scrollengine/engine_navigation.cpp
+    src/scrollengine/engine_verbs.cpp
     src/scrollengine/engine_serialize.cpp
     src/scrollengine/scrollenginelogging.h
     src/scrollengine/scrollenginelogging.cpp
