+++ libs/phosphor-scroll-engine/src/scrollengine/engine_navigation.cpp
@@ -757,151 +757,10 @@ void ScrollEngine::toggleFocusedFloatAs(const PhosphorEngine::NavigationContext&
     toggleWindowFloat(windowId, ctx.screenId);
 }
 
-// ── Scroll-specific vocabulary ──────────────────────────────────────────────
+// ── Scroll-specific vocabulary lives in engine_verbs.cpp ────────────────────
+// The P_SCROLL_VERB one-liner family (focus/move/consume/center/width/height
+// verbs) moved there wholesale; P_SCROLL_RESOLVE is duplicated in that TU.
 
-// Body shared by every parameterless column verb: run the strip op, then
-// relayout + activate + notify when it changed something.
-#define P_SCROLL_VERB(screenIdExpr, opExpr, actionStr)                                                                 \
-    P_SCROLL_RESOLVE(screenIdExpr);                                                                                    \
-    if (!state || state->strip().isEmpty()) {                                                                          \
-        Q_EMIT navigationFeedback(false, QStringLiteral(actionStr), QStringLiteral("no_windows"), QString(),           \
-                                  QString(), screen);                                                                  \
-        return;                                                                                                        \
-    }                                                                                                                  \
-    const QString sourceWindow = state->strip().activeWindowId();                                                      \
-    const bool changed = (opExpr);                                                                                     \
-    if (changed) {                                                                                                     \
-        applyLayout(screen, true);                                                                                     \
-        Q_EMIT placementChanged(screen);                                                                               \
-    }                                                                                                                  \
-    Q_EMIT navigationFeedback(changed, QStringLiteral(actionStr), changed ? QString() : QStringLiteral("no_target"),   \
-                              sourceWindow, changed ? state->strip().activeWindowId() : QString(), screen)
-
-void ScrollEngine::focusColumnFirst(const QString& screenId)
-{
-    P_SCROLL_VERB(screenId, state->strip().focusFirstColumn(params), "focus");
-}
-
-void ScrollEngine::focusColumnLast(const QString& screenId)
-{
-    P_SCROLL_VERB(screenId, state->strip().focusLastColumn(params), "focus");
-}
-
-void ScrollEngine::moveColumnToFirst(const QString& screenId)
-{
-    P_SCROLL_VERB(screenId, state->strip().moveActiveColumnToFirst(params), "move");
-}
-
-void ScrollEngine::moveColumnToLast(const QString& screenId)
-{
-    P_SCROLL_VERB(screenId, state->strip().moveActiveColumnToLast(params), "move");
-}
-
-// NOTE on the P_SCROLL_* macros above: they deliberately inject `screen`,
-// `state`, and `params` into the caller's scope and embed an early return.
-// A helper struct + lambda was considered and rejected: every verb would
-// still need the three names plus the bail-out, and the macro keeps 14 of the
-// 16 verb bodies one line each (toggleColumnTabbed and resetWindowHeights are
-// hand-expanded — neither op reads layout params). The names are part of the
-// macro's documented contract, and both macros are #undef'd at the end of this
-// file.
-void ScrollEngine::consumeWindowIntoColumn(const QString& screenId)
-{
-    P_SCROLL_VERB(screenId, state->strip().consumeWindowIntoColumn(params), "consume");
-}
-
-void ScrollEngine::expelWindowFromColumn(const QString& screenId)
-{
-    P_SCROLL_VERB(screenId, state->strip().expelWindowFromColumn(params), "expel");
-}
-
-void ScrollEngine::consumeOrExpelWindow(int delta, const QString& screenId)
-{
-    P_SCROLL_VERB(screenId, state->strip().consumeOrExpel(delta, params), "consume");
-}
-
-void ScrollEngine::centerColumn(const QString& screenId)
-{
-    P_SCROLL_VERB(screenId, state->strip().centerActiveColumn(params), "center");
-}
-
-void ScrollEngine::toggleColumnTabbed(const QString& screenId)
-{
-    // Hand-expanded (not P_SCROLL_VERB): the op never reads layout params,
-    // and the macro's resolve pays a ScreenManager query plus a
-    // context-gap-provider invocation per call — same reasoning as
-    // snapAllWindows.
-    const QString screen = resolveOperationScreen(screenId);
-    ScrollState* state = screen.isEmpty() ? nullptr : stateForKey(currentKeyForScreen(screen), false);
-    if (!state || state->strip().isEmpty()) {
-        Q_EMIT navigationFeedback(false, QStringLiteral("tabbed"), QStringLiteral("no_windows"), QString(), QString(),
-                                  screen);
-        return;
-    }
-    const QString sourceWindow = state->strip().activeWindowId();
-    const bool changed = state->strip().toggleActiveColumnTabbed();
-    if (changed) {
-        applyLayout(screen, true);
-        Q_EMIT placementChanged(screen);
-    }
-    Q_EMIT navigationFeedback(changed, QStringLiteral("tabbed"), changed ? QString() : QStringLiteral("no_target"),
-                              sourceWindow, changed ? state->strip().activeWindowId() : QString(), screen);
-}
-
-void ScrollEngine::cycleColumnPresetWidth(int delta, const QString& screenId)
-{
-    P_SCROLL_VERB(screenId, state->strip().cycleActiveColumnPresetWidth(delta, params), "resize");
-}
-
-void ScrollEngine::adjustColumnWidth(qreal deltaPercent, const QString& screenId)
-{
-    P_SCROLL_VERB(screenId, state->strip().adjustActiveColumnWidth(deltaPercent, params), "resize");
-}
-
-void ScrollEngine::toggleMaximizeColumn(const QString& screenId)
-{
-    P_SCROLL_VERB(screenId, state->strip().toggleMaximizeActiveColumn(params), "resize");
-}
-
-void ScrollEngine::expandColumnToAvailableWidth(const QString& screenId)
-{
-    P_SCROLL_VERB(screenId, state->strip().expandActiveColumnToAvailableWidth(params), "resize");
-}
-
-void ScrollEngine::cycleWindowPresetHeight(int delta, const QString& screenId)
-{
-    P_SCROLL_VERB(screenId, state->strip().cycleActiveWindowPresetHeight(delta, params), "resize");
-}
-
-void ScrollEngine::adjustWindowHeight(qreal deltaPercent, const QString& screenId)
-{
-    P_SCROLL_VERB(screenId, state->strip().adjustActiveWindowHeight(deltaPercent, params), "resize");
-}
-
-void ScrollEngine::resetWindowHeights(const QString& screenId)
-{
-    // Hand-expanded (not P_SCROLL_VERB): the op never reads layout params,
-    // and the macro's resolve pays a ScreenManager query plus a
-    // context-gap-provider invocation per call — same reasoning as
-    // snapAllWindows.
-    const QString screen = resolveOperationScreen(screenId);
-    ScrollState* state = screen.isEmpty() ? nullptr : stateForKey(currentKeyForScreen(screen), false);
-    if (!state || state->strip().isEmpty()) {
-        Q_EMIT navigationFeedback(false, QStringLiteral("resize"), QStringLiteral("no_windows"), QString(), QString(),
-                                  screen);
-        return;
-    }
-    const QString sourceWindow = state->strip().activeWindowId();
-    const bool changed = state->strip().resetActiveColumnHeights();
-    if (changed) {
-        applyLayout(screen, true);
-        Q_EMIT placementChanged(screen);
-    }
-    Q_EMIT navigationFeedback(changed, QStringLiteral("resize"), changed ? QString() : QStringLiteral("no_target"),
-                              sourceWindow, changed ? state->strip().activeWindowId() : QString(), screen);
-}
-
-#undef P_SCROLL_VERB
 #undef P_SCROLL_RESOLVE
 
 } // namespace PhosphorScrollEngine
